[CircleGraph] Introduce CircleGraphCtrl
This will extract part of CircleGraphPanel to as CircleGraphCtrl to
provide common CircleGraph as a UI component.
- this change will extract getHtmlForWebview and related

ONE-vscode-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
