+++ src/CircleGraph/CircleGraphCtrl.ts
@@ -1,0 +1,95 @@
+/*
+ * Copyright (c) 2022 Samsung Electronics Co., Ltd. All Rights Reserved
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import * as fs from 'fs';
+import * as vscode from 'vscode';
+
+export class CircleGraphCtrl {
+  // TODO make this protected
+  public static readonly folderMediaCircleGraph = 'media/CircleGraph';
+  protected static readonly folderMediaCircleGraphExt = 'media/CircleGraph/external';
+  protected static readonly folderExternal = 'external/';
+
+  protected readonly _extensionUri: vscode.Uri;
+
+  public constructor(extensionUri: vscode.Uri) {
+    this._extensionUri = extensionUri;
+  }
+
+  public getHtmlForWebview(webview: vscode.Webview) {
+    const htmlPath = this.getMediaPath('index.html');
+    let html = fs.readFileSync(htmlPath.fsPath, {encoding: 'utf-8'});
+
+    const nonce = getNonce();
+    html = html.replace(/\%nonce%/gi, nonce);
+    html = html.replace('%webview.cspSource%', webview.cspSource);
+    // necessary files from netron to work
+    html = this.updateExternalUri(html, webview, '%view-grapher.css%', 'view-grapher.css');
+    html = this.updateExternalUri(html, webview, '%view-sidebar.css%', 'view-sidebar.css');
+    html = this.updateExternalUri(html, webview, '%view-sidebar.js%', 'view-sidebar.js');
+    html = this.updateExternalUri(html, webview, '%view-grapher.js%', 'view-grapher.js');
+    html = this.updateExternalUri(html, webview, '%dagre.js%', 'dagre.js');
+    html = this.updateExternalUri(html, webview, '%base.js%', 'base.js');
+    html = this.updateExternalUri(html, webview, '%text.js%', 'text.js');
+    html = this.updateExternalUri(html, webview, '%json.js%', 'json.js');
+    html = this.updateExternalUri(html, webview, '%xml.js%', 'xml.js');
+    html = this.updateExternalUri(html, webview, '%python.js%', 'python.js');
+    html = this.updateExternalUri(html, webview, '%protobuf.js%', 'protobuf.js');
+    html = this.updateExternalUri(html, webview, '%flatbuffers.js%', 'flatbuffers.js');
+    html = this.updateExternalUri(html, webview, '%flexbuffers.js%', 'flexbuffers.js');
+    html = this.updateExternalUri(html, webview, '%zip.js%', 'zip.js');
+    html = this.updateExternalUri(html, webview, '%gzip.js%', 'gzip.js');
+    html = this.updateExternalUri(html, webview, '%tar.js%', 'tar.js');
+    // for circle format
+    html = this.updateExternalUri(html, webview, '%circle.js%', 'circle.js');
+    html = this.updateExternalUri(html, webview, '%circle-schema.js%', 'circle-schema.js');
+    // modified for one-vscode
+    html = this.updateUri(html, webview, '%index.js%', 'index.js');
+    html = this.updateUri(html, webview, '%view.js%', 'view.js');
+
+    return html;
+  }
+
+  private getMediaPath(file: string) {
+    return vscode.Uri.joinPath(this._extensionUri, CircleGraphCtrl.folderMediaCircleGraph, file);
+  }
+
+  private updateExternalUri(
+      html: string, webview: vscode.Webview, search: string, replace: string) {
+    const replaceUri = this.getUriFromPath(webview, CircleGraphCtrl.folderExternal + replace);
+    return html.replace(search, `${replaceUri}`);
+  }
+
+  private updateUri(html: string, webview: vscode.Webview, search: string, replace: string) {
+    const replaceUri = this.getUriFromPath(webview, replace);
+    return html.replace(search, `${replaceUri}`);
+  }
+
+  private getUriFromPath(webview: vscode.Webview, file: string) {
+    const mediaPath = this.getMediaPath(file);
+    const uriView = webview.asWebviewUri(mediaPath);
+    return uriView;
+  }
+};
+
+function getNonce() {
+  let text = '';
+  const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  for (let i = 0; i < 32; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+  return text;
+}
